simplifying serializer class
